Adjust config tool pas range (#592)
* added new range to some PAS parameters

* modified pas detection ranges

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/application/user_config_data.h
@@ -82,7 +82,7 @@ typedef struct
 {
     uint8_t pasTorqueStartupSpeed;
     uint8_t pasTorqueStartupThreshold;    
-    uint8_t pasCadenceStartupNumbPulses;
+    uint16_t pasCadenceStartupNumbPulses;
     uint16_t pasCadenceStartupWindows;
     uint8_t PasAlgorithmStartup;
 } PAS_Startup_Detection_t;
@@ -90,7 +90,7 @@ typedef struct
 typedef struct
 {
     uint8_t pasTorqueRunningThreshold;    
-    uint8_t pasCadenceRunningNumbPulses;
+    uint16_t pasCadenceRunningNumbPulses;
     uint16_t pasCadenceRunningWindows;
     uint8_t PasAlgorithmRunning;
 } PAS_Running_Detection_t;
